requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosqlite==0.21.0
attrs==25.4.0
black==25.9.0
click==8.3.0
defusedxml==0.7.1
dnspython==2.8.0
frozenlist==1.8.0
idna==3.11
ipwhois==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-whois==0.9.6
pytokens==0.2.0
six==1.17.0
typing_extensions==4.15.0
yarl==1.22.0
